Add R code to generate git commit information
It will help us juggle multiple versions of the paper (for editting, etc)
if the git commit information and date is embedded in the final pdf.
To make sure that we have a cross-platform method of generating that
commit information, I'm using the `git2r` R library to generate it.
The R script in this commit makes a file `VERSION.tex` that provides
the \VERSION command containing the git commit information.

\VERSION can be overwritten by defining it before importing `VERSION.tex`.