Add main R script to manage code execution
This project will have several R scripts: some that generate the Monte
Carlo results; some to do an empirical study; some to grab and display
information about the Git repository; etc. So this commit adds a top
level R script that makes sure they all run. We will probably want to
tweak it so that the results are cached, but for now they just run
straight through.

The commit also creates an R subdirectory that will contain all of the
other R scripts (not `run_everything.R` though) and moves the existing
simulation code there.