Add fallback when VERSION.tex
Some of the people working on this project (ahem) are not as
excited about the idea of 'automatically generated git commit
information' as others and have resisted generating VERSION.tex.
This commit makes an ugly but seemingly necesary concession to
them and adds a default fallback option if VERSION.tex does not
exist. You're mfing welcome.