Rewrite second mathematical result and start its proof
This commit rewrites the second mathematical result and restates
it as a theorem that the 'double bootstrap' procedure accurately
approximates the asymptotic distribution of the bias-corrected
statistic we propose in Theorem and Algorithm 1.

This commit also has the first draft of a proof of the second
theorem as well as a crucial supporting lemma that actually
proves normality. There are still some steps missing and some
of the scaling factors are not yet correct. So there's work to
be done!