Create R script to run all analysis
This commit adds (back) the function R/run_everything.R, which runs
all of the simulations and empirical analysis for the paper. It also
moves the R code that replicates Ludwig and Miller's tables into a
separate file to keep the application code as simple as possible. Each
R script is run in a separate environment to explicitly keep the
workspace clean.