Tweak notation for proof of first theorem
Our notation didn't distinguish between the estimate of the first derivative
of the conditional expectation using a 1st order vs 2nd order expansion. I've
updated the paper to include that distinction. It changes the notation definitions
at the start of the appendix as well as the proof of the first theorem.