First full draft of slides; they still have known errors
This commit adds the source code for the first draft of slides for
the "Advances in Econometrics" conference on RDD. The slides are
bad and have known errors. But they have all the material, so I'm
checking them in.

The source for the slides is in an org-mode file. They're exported
to pdf by creating an intermediate LaTeX file. If you need to
work with the LaTeX file, let me (Gray) know and I'll help
produce it.