[#3] Fix title page for RDD conference slides
Previously, Gray Calhoun was the only one listed on the title page for the slides. This
commit adds Otavio and Yang. I'm not thrilled with the spacing, but the org-mode
exporter is a little finicky. I may tweak it some more later, but everyone is listed now
so that's a major improvement.