Refactor bootstrap functions
This commit rewrites the bootstrap code so that it is more modular and
easier to modify. The algorithm is the same for now, but the random
variables produced by the RNG are used in a different order than before,
so it will not produce numerically identical results to the earlier
version except in some special circumstances. I've also added a test
file that verifies that the new output matches the old output in some
of those settings.