Make small refinement to Theorem 2
We should be clear that the distributional convergence in Theorem 2 is
uniform across tau. That's not necessary in Theorem 1 because the limit
is normal so convergence is automatically uniform across different
values of the mean. In Theorem 2 we're not explicit that the limit is
normal so we should be explicit about uniformity.